Move types parser.Data and parserdata.Location to new module basic.
